+++ src/simple/internal/fuzzers/strsim/metrics/jaro.rs
@@ -12,7 +12,6 @@ pub struct Jaro;
 impl crate::simple::internal::fuzzers::strsim::Metric for Jaro {
     /// Similarity metric. Inversion of relative distance, ranging from 1.0
     /// (equality) to 0.0 (nothing in common).
-    #[must_use]
     #[inline(always)]
     fn similarity(str1: &str, str2: &str) -> f64 {
         strsim::jaro(str1, str2)