+++ src/simple/internal/fuzzers/strsim/metrics/jaro_winkler.rs
@@ -13,7 +13,6 @@ pub struct JaroWinkler;
 impl crate::simple::internal::fuzzers::strsim::Metric for JaroWinkler {
     /// Similarity metric. Inversion of relative distance, ranging from 1.0
     /// (equality) to 0.0 (nothing in common).
-    #[must_use]
     #[inline(always)]
     fn similarity(str1: &str, str2: &str) -> f64 {
         strsim::jaro_winkler(str1, str2)